refactor: 코스 목록 카드 관련 코드 리팩토링
