+++ src/features/course-card/index.ts
@@ -1,0 +1,2 @@
+export * from './grid-card'
+export * from './list-card'
