--- src/features/course-plan-card/index.ts
@@ -1,1 +1,0 @@
-export * from './ui'
