--- src/features/course-plan-card/ui/course-grid-card.tsx
@@ -1,78 +1,0 @@
-import Link from 'next/link'
-import Image from 'next/image'
-import ProfileImage from '@/src/shared/ui/ProfileImage'
-import { Heart, MessageSquare, Share2 } from 'lucide-react'
-import { CourseType } from '@/src/entities/course'
-import logo from '@/src/assets/images/(logo)/logo.png'
-
-export function CourseGridCard({ course }: { course: CourseType }) {
-  const handleClickLike = () => {}
-
-  return (
-    <div className='w-[164px] h-[217px] flex flex-col justify-between pb-[8px] rounded-[10px] bg-white drop-shadow-[0_0_4px_rgba(0,0,0,0.15)]'>
-      <Link
-        href={`/courses/${course.id}`}
-        className='w-full flex flex-col gap-[10px]'
-      >
-        <Image
-          src={course.places[0].thumbnail_url || logo}
-          width={207}
-          height={100}
-          className='h-[100px] bg-light-gray rounded-tr-[10px] rounded-tl-[10px] object-cover'
-          alt='course-image'
-        />
-
-        <section className='absolute top-[82px] left-[10px] w-[27px] h-[27px] bg-gradient-to-r from-[#9997F2] to-[#4341EA] p-[1px] rounded-[50%]'>
-          <ProfileImage
-            size={25}
-            src={course.writer.profile_url}
-            userId={course.writer.id}
-          />
-        </section>
-
-        <section className='flex flex-col gap-[7px] h-full px-[11px] mt-[3px] leading-none'>
-          <span className='text-sub text-brand font-medium h-[11px]'>
-            {course.writer.name}
-          </span>
-          <div className='flex flex-row items-center justify-between'>
-            <span className='text-sub text-black font-light'>
-              {course.primary_region} / {course.secondary_region}
-            </span>
-
-            <span className='text-[9px] text-search-gray opacity-80 font-light'>
-              {course.visit_date}
-            </span>
-          </div>
-          <span className='text-[13px] font-semibold w-full text-words line-clamp-2 leading-4'>
-            {course.title}
-          </span>
-        </section>
-      </Link>
-
-      <section className='flex items-center justify-between text-sub text-description px-[11px]'>
-        <div className='flex flex-row items-center gap-[9px]'>
-          <div className='flex items-center gap-[4px]'>
-            <Heart
-              size={15}
-              className='text-brand'
-              fill={course.is_liked ? '#5A59F2' : 'none'}
-              strokeWidth={1.5}
-              onClick={handleClickLike}
-            />
-            <span>{course.likes}</span>
-          </div>
-          <div className='flex items-center gap-[4px]'>
-            <MessageSquare size={15} className='text-brand' strokeWidth={1.5} />
-            <span>{course.comments}</span>
-          </div>
-        </div>
-        <Share2
-          size={16}
-          className='cursor-pointer text-brand'
-          strokeWidth={1.5}
-          onClick={() => {}}
-        />
-      </section>
-    </div>
-  )
-}
