--- src/features/course-plan-card/ui/index.ts
@@ -1,3 +1,0 @@
-export * from './default-card'
-export * from './course-grid-card'
-export * from './course-list-card'
