+++ src/features/index.ts
@@ -1,7 +1,8 @@
-export * from './course-plan-card'
+export * from './course-action-bar'
 export * from './place-collapse'
 export * from './place-card'
 export * from './place-review-stats'
 export * from './review-tag'
 export * from './tabs'
 export * from './select-sort'
+export * from './course-card'
