+++ src/widgets/course-plan-card/course-plan-card.tsx
